+++ cli/nep_test/nep17_test.go
@@ -1,4 +1,4 @@
-package wallet_test
+package nep_test
 
 import (
 	"io"
